+++ src/main/kotlin/us/frollo/frollosdk/FrolloSDK.kt
@@ -26,6 +26,7 @@ import org.threeten.bp.Duration
 import org.threeten.bp.LocalDate
 import org.threeten.bp.LocalDateTime
 import us.frollo.frollosdk.address.AddressManagement
+import us.frollo.frollosdk.affordability.Affordability
 import us.frollo.frollosdk.aggregation.Aggregation
 import us.frollo.frollosdk.authentication.AuthenticationStatus
 import us.frollo.frollosdk.authentication.AuthenticationType.Custom
@@ -224,6 +225,12 @@ object FrolloSDK {
     val serviceStatusManagement: ServiceStatusManagement
         get() = _serviceStatusManagement ?: throw IllegalAccessException(SDK_NOT_SETUP)
 
+    /**
+     * Affordability - Affordability management. See [Affordability] for details
+     */
+    val affordability: Affordability
+        get() = _affordability ?: throw IllegalAccessException(SDK_NOT_SETUP)
+
     private var _setup = false
     private var _logger: LogManager? = null
     private var _aggregation: Aggregation? = null
@@ -246,6 +253,7 @@ object FrolloSDK {
     private var _statements: Statements? = null
     private var _addressManagement: AddressManagement? = null
     private var _serviceStatusManagement: ServiceStatusManagement? = null
+    private var _affordability: Affordability? = null
     private lateinit var keyStore: Keystore
     private lateinit var preferences: Preferences
     private lateinit var version: Version
@@ -395,6 +403,9 @@ object FrolloSDK {
 
             // 29. Setup Service Status Management
             _serviceStatusManagement = ServiceStatusManagement(network, database)
+
+            // 30. Financial passport management
+            _affordability = Affordability(network)
 
             if (version.migrationNeeded()) {
                 version.migrateVersion()